services/email/requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
pika==1.3.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
resend==2.4.0
typing_extensions==4.12.2
urllib3==2.2.3
